Fix broken reset onboarding behavior on home page
