+++ src/routes/Onboarding/index.tsx
@@ -20,7 +20,6 @@ import { useAbsoluteFilePath } from 'hooks/useAbsoluteFilePath'
 import { ActionButton } from 'components/ActionButton'
 import { onboardingPaths } from 'routes/Onboarding/paths'
 
-export const ONBOARDING_PROJECT_NAME = 'Tutorial Project $nn'
 export const kbdClasses =
   'p-0.5 text-sm rounded-sm bg-chalkboard-10 dark:bg-chalkboard-100 border border-chalkboard-50'
